Add heap and stack to test (#245)
* Add test_stack io tests

* Add test_heap to test
