+++ stack/simplify_path.py
@@ -10,8 +10,6 @@ path = "/a/./b/../../c/", => "/c"
 * Another corner case is the path might contain multiple slashes '/' together, such as "/home//foo/".
     In this case, you should ignore redundant slashes and return "/home/foo".
 """
-
-
 def simplify_path(path):
     """
     :type path: str
@@ -27,10 +25,3 @@ def simplify_path(path):
         elif tok not in skip:
             stack.append(tok)
     return '/' + '/'.join(stack)
-
-
-if __name__ == '__main__':
-
-    p = '/my/name/is/..//keon'
-    print(p)
-    print(simplify_path(p))
